test(ios-app): gate the composer-under-trouble UI tests in CI
`ios-app-tests` runs only `SliccFollowerTests` on the iPhone legs and two
named classes on the iPad ones, so UI tests added to
`ComposerKeyboardUITests` would never have executed in CI — a regression
guard that never runs guards nothing.

Move them to their own `ComposerConnectionUITests` and run that class on
the iPhone/26 leg. A separate class rather than the whole existing one:
`ComposerKeyboardUITests` drives `typeKey`, which needs a hardware
keyboard the runner does not have, so those two tests fail wherever they
run (including on `main`) and must stay out of the gate.
